+++ src/components/FileExplorer.vue
@@ -42,11 +42,10 @@
         </div>
         <span class="col-grow" />
         <div class="row no-wrap">
-          <file-explorer-action-card
-            class="file-explorer-button"
+          <file-explorer-action-button
+            class="file-explorer-action-button"
             :file="node"
             :project-name="projectName"
-            :data-cy="`${node.isFolder ? 'folder': 'file'}-button_${node.id}`"
           />
         </div>
       </div>
@@ -64,7 +63,7 @@ import {
   toRef,
 } from 'vue';
 import GitEvent from 'src/composables/events/GitEvent';
-import FileExplorerActionCard from 'src/components/card/FileExplorerActionCard.vue';
+import FileExplorerActionButton from 'src/components/FileExplorerActionButton.vue';
 import FileName from 'src/components/FileName.vue';
 import { getTree, updateFileInformation } from 'src/composables/FileExplorer';
 import { getProjectFiles, getStatus } from 'src/composables/Project';
@@ -96,6 +95,7 @@ const filterTrigger = ref(toRef(props, 'showParsableFiles').value.toString());
 
 let selectFileTabSubscription;
 let createFileSubscription;
+let renameFileSubscription;
 let deleteFileSubscription;
 let updateEditorContentSubscription;
 let addRemoteSubscription;
@@ -328,9 +328,10 @@ async function openModelFiles() {
 /**
  * Update localFileInformations and nodes of the tree then update all status.
  * Also, expand model folders and open parsable files.
+ * @param {boolean} avoidOpening - Avoid automatic model file(s) opening.
  * @returns {Promise<void>} Promise with nothing on success otherwise an error.
  */
-async function initTreeNodes() {
+async function initTreeNodes(avoidOpening = false) {
   localFileInformations.value = await getProjectFiles(props.projectName);
 
   nodes.value = getTree(props.projectName, localFileInformations.value);
@@ -338,7 +339,7 @@ async function initTreeNodes() {
   await updateAllFilesStatus();
 
   // TODO: Find a better way to stub it on shallowMount.
-  if (fileExplorerRef.value.getNodeByKey) {
+  if (fileExplorerRef.value.getNodeByKey && !avoidOpening) {
     openModelFiles();
   }
 }
@@ -349,6 +350,9 @@ onMounted(async () => {
   });
   createFileSubscription = FileEvent.CreateFileEvent.subscribe((event) => {
     onCreateFile(event);
+  });
+  renameFileSubscription = FileEvent.RenameFileEvent.subscribe(() => {
+    initTreeNodes(true);
   });
   deleteFileSubscription = FileEvent.DeleteFileEvent.subscribe(onDeleteFile);
   updateEditorContentSubscription = FileEvent.UpdateEditorContentEvent.subscribe((event) => {
@@ -379,6 +383,7 @@ onMounted(async () => {
 onUnmounted(() => {
   selectFileTabSubscription.unsubscribe();
   createFileSubscription.unsubscribe();
+  renameFileSubscription.unsubscribe();
   deleteFileSubscription.unsubscribe();
   updateEditorContentSubscription.unsubscribe();
   addRemoteSubscription.unsubscribe();
@@ -411,12 +416,12 @@ onUnmounted(() => {
     background: rgba($primary, 0.1);
     border-radius: 4px;
 
-    .file-explorer-button {
+    .file-explorer-action-button {
       display: inline-block
     }
   }
 }
-.file-explorer-button {
+.file-explorer-action-button {
   display: none
 }
 </style>