+++ changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Hide component detail panel when deleting corresponding component.
 * Autosave on component detail panel.
 * Debounce of 1 second on file auto-save.
+* Add 'Rename' action in FileExplorer file/folder menu.
 * Setup conditional features:
   * `CF_createDiagram`: allows or prevents user to create a diagram.
   * `CF_createDiagramFromTemplate`: allows or prevents user to create a diagram from a template.